SMACK-389 Plugable compression framework
This commit adds the ability to plug-in different compression handlers
for different compression methods. It is also possible to add more
then one handler for the same method. The order how the handlers are
added determines which handler is tried first for usage. This is used
for the 'zlib' compression method, which now can be either provided by
Java7 or by JZlib (just like it was done before).

git-svn-id: http://svn.igniterealtime.org/svn/repos/smack/trunk@13522 b35dd754-fafc-0310-a699-88a17e54d16e

Conflicts:
	build/eclipse/classpath
	source/org/jivesoftware/smack/Connection.java
	source/org/jivesoftware/smack/XMPPConnection.java
	source/org/jivesoftware/smack/compression/Java7ZlibInputOutputStream.java
	source/org/jivesoftware/smack/compression/JzlibInputOutputStream.java
	source/org/jivesoftware/smack/compression/XMPPInputOutputStream.java